# src/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the API that serves as the central interaction point between the client and back-end services of Earthdata Pub.",
    "version": "1.0.0",
    "title": "EarthdataPub",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brian.ellingson@uah.edu"
    }
  },
...
  "paths": {
...
    "/api/token": {
      "get": {
        "tags": [
          "token"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from identity provider.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "A redirect url to redirect to after successful authentication.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "logout",
            "in": "query",
            "description": "Set to true to get redirect uri to logout through identity provider.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "summary": "Authentication and token management",
        "description": "Handles authentication and token management with an identity provider.",
        "operationId": "getToken",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "${auth_lambda_arn}",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "POST",
          "type": "aws",
          "requestTemplates": {
            "application/json": "#set($req=$input.params()) {#foreach($field in $req.querystring.keySet()) \"$field\":\"$req.querystring.get($field)\" #if($foreach.hasNext),#end #end  }"
          },
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'*'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key'"
              }
            }
          }
        },
        "x-router-controller": "proxy"
      },
      "options": {
        "tags": [
          "cors"
        ],
        "summary": "CORS",
        "description": "OPTIONS method for CORS",
        "operationId": "optionsToken",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200 }"
          },
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'*'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key'"
              },
              "responseTemplates": {
                "application/json": "{}"
              }
            }
          }
        },
        "x-router-controller": "cors"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Cognito_Authorizer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "This authorizer is used for most requests to the EDPub API with the exception of service endpoints. The authorization value follows the format: 'Bearer \\<JWT token value\\>'",
        "x-amazon-apigateway-authtype": "cognito_user_pools",
        "x-amazon-apigateway-authorizer": {
          "type": "cognito_user_pools",
          "providerARNs": [ "${cognito_user_pool_arn}" ]
        }
      },
      "Service_Authorizer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "This authorizer is used for authentication for service enabled endpoints. The authorization value format is: 'Basic \\<value of encoding serviceId:serviceSecret using base64\\>'",
        "x-amazon-apigateway-authtype" : "custom",
        "x-amazon-apigateway-authorizer" : {
          "type" : "request",
          "authorizerUri" : "${service_authorizer_lambda_arn}",
          "authorizerPayloadFormatVersion" : "1.0",
          "identitySource" : "method.request.header.authorization, method.request.header.submissionid",
          "authorizerResultTtlInSeconds" : "300"
        }
      }
    },
...
  },
...
}